backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
sqlparse==0.4.4
openai==1.3.5
anthropic==0.7.7
pydantic==2.5.0
python-json-logger==2.0.7
uvicorn==0.24.0
websockets==12.0